+++ src/app/shared/helpers/date-picker.tns.ts
@@ -2,7 +2,7 @@ import { ModalDatetimepicker } from 'nativescript-modal-datetimepicker';
 
 import { dateToString } from '../misc';
 
-export function openDatePicker(): Promise<string> {
+export function openDatePicker(datepicker: void): Promise<string> { // eslint-disable-line @typescript-eslint/no-unused-vars
     const picker = new ModalDatetimepicker();
     return picker.pickDate({
         theme: 'overlay',